+++ PaperDetailTableViewController.m
@@ -192,7 +192,7 @@
 					NSUInteger count = words.count;
 					if (((NSString *)[words objectAtIndex:0]).length == 0) --count;
 					if (count > 0 && ((NSString *)[words lastObject]).length == 0) --count;
-					((UITextField *)wordCountCell.accessoryView).text = [NSString stringWithFormat:@"%ld", count];
+					((UITextField *)wordCountCell.accessoryView).text = [NSString stringWithFormat:@"%ld", (unsigned long)count];
 					return wordCountCell;
 					break;
 				case rowAPAInstitution:
